compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018998
#> Epoch: Iteration #200 error is: 217.144151875719
#> Epoch: Iteration #300 error is: 217.262471696479
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  12.2361793   4.2088494  -4.6712543  17.9965545   8.8794348 -10.9998721
#>  [7] -26.5628045  -6.9469239 -30.3256428 -20.7983135 -14.9411080 -23.5926509
#> [13]  15.4799753  20.5049708  30.9075101   1.9802919  27.1487917  -2.3293618
#> [19]  -0.2436301 -17.3970657   5.8807585  23.5577640  -8.9799536 -13.0699847
#> [25]   7.2365786  10.4073359  13.4604784 -19.0269072
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018114
#> Epoch: Iteration #200 error is: 217.143553779776
#> Epoch: Iteration #300 error is: 217.276009048734

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018033
#> Epoch: Iteration #200 error is: 217.136609517831
#> Epoch: Iteration #300 error is: 217.269267325644
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  12.2361793   4.2088494  -4.6712543  17.9965545   8.8794348 -10.9998721
#>  [7] -26.5628045  -6.9469239 -30.3256428 -20.7983135 -14.9411080 -23.5926509
#> [13]  15.4799753  20.5049708  30.9075101   1.9802919  27.1487917  -2.3293618
#> [19]  -0.2436301 -17.3970657   5.8807585  23.5577640  -8.9799536 -13.0699847
#> [25]   7.2365786  10.4073359  13.4604784 -19.0269072
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1]  15.0373189   4.0217663  -4.6311868  17.6523127  10.0134227 -10.8790656
#>  [7] -29.0284533  -6.8849730 -20.0918899 -22.1317831 -14.7251338 -25.4314052
#> [13]  12.6798057  20.1390147  30.4291814   1.9145187  26.7012095  -2.3159168
#> [19]  -0.2738968 -17.1523181   7.2265451  23.1779778  -8.8964319 -12.9382667
#> [25]   5.9398038   8.3353480  10.9252720 -18.8127765