compute network diversity indices

Function to compute network diversity indices based on Hill numbers following the method described in Ohlmann et al. 2019

compute_diversities(metanetwork, q = 1, res = NULL)

Arguments

metanetwork

object of class 'metanetwork'

q

viewpoint parameter controlling the weight given to abundant species/groups and links

res

a vector containing the resolutions at which the diversities are computed

Value

a data.frame

Details

XXXXXXXXXXXXXXXXXXXXXx

References

Ohlmann, M., Miele, V., Dray, S., Chalmandrier, L., O'connor, L., & Thuiller, W. (2019). Diversity indices for ecological networks: a unifying framework using Hill numbers. Ecology letters, 22(4), 737-747.

Examples

library(metanetwork)
library(igraph)

#on angola dataset
data("meta_angola")
compute_diversities(meta_angola,q = 1)
#> $nodes
#>                 Species   Phylum
#> Gamma_P       13.666059 2.680737
#> mean_Alpha_P  12.478328 2.669698
#> Beta_P         1.095183 1.004135
#> Alpha_X1986_P  9.027714 2.494158
#> Alpha_X2003_P 17.247852 2.857593
#> 
#> $links
#>                 Species   Phylum
#> Gamma_L       55.718125 3.812301
#> mean_Alpha_L  48.505518 3.791652
#> Beta_L         1.148697 1.005446
#> Alpha_X1986_L 36.481408 3.410505
#> Alpha_X2003_L 64.492722 4.215394
#> 

#computing diversities only at Phylum level
compute_diversities(meta_angola,q = 1,res = "Phylum")
#> $nodes
#>                 Phylum
#> Gamma_P       2.680737
#> mean_Alpha_P  2.669698
#> Beta_P        1.004135
#> Alpha_X1986_P 2.494158
#> Alpha_X2003_P 2.857593
#> 
#> $links
#>                 Phylum
#> Gamma_L       3.812301
#> mean_Alpha_L  3.791652
#> Beta_L        1.005446
#> Alpha_X1986_L 3.410505
#> Alpha_X2003_L 4.215394
#>