Runs Consensus Clustering Algorithm

A modiefed parallel version of code imported from https://github.com/Bioconductor-mirror/ConsensusClusterPlus 1.11.1. Taken out of findModules.consensusCluster.R

run.consensus.cluster(
  d,
  kGrid = NULL,
  repCount = NULL,
  diss = inherits(d, "dist"),
  pItem = NULL,
  pFeature = NULL,
  innerLinkage = NULL,
  distance = NULL,
  clusterAlg = NULL,
  weightsItem = NULL,
  weightsFeature = NULL,
  verbose = NULL,
  corUse = NULL
)

Arguments

d

Required. A matrix where columns=items/samples and rows are features. For example, a gene expression matrix of genes in rows and microarrays in columns. OR ExpressionSet object. (Default = NULL)

kGrid

Optional. (Default = NULL)

repCount

Optional. Replicate count. (Default = NULL)

diss

Optional. A distance matrix of d. (Default = inherits( d, "dist" ))

pItem

Optional. A numerical value. proportion of items to sample. (Default = 0.8)

pFeature

Optional. A numerical value. proportion of features to sample. (Default = 1)

innerLinkage

Optional. A heirarchical linkage method for subsampling. (Default = "average")

distance

Optional. A character value. sample distance measures: "pearson","spearman", or "euclidean". (Default = "pearson")

clusterAlg

Optional. A character value. cluster algorithm. "hc" heirarchical (hclust) or "km" for kmeans. (Default = "kmeans")

weightsItem

Optional. A numerical vector. weights to be used for sampling items. (Default = NULL)

weightsFeature

Optional. AN umerical vector. weights to be used for sampling features. (Default = NULL)

verbose

Optional. A boolean when set to TRUE, prints messages to the screen to indicate progress. This is useful for large datasets.(Default = FALSE)

corUse

Optional. Use all cores avaiable. (Default = "Everything")

Value

Clustered consensus matrix. areaUnderCDF = areaK, consensus.matrix = cns.mtrx