Simulate Network Data

This function simulates a gene correlation network and inverse covariance matrix of user determined size.

simulateNetworkData(n, p, prop, adjustment = 10)

Arguments

n

Required. The number of rows to simulate in the data. These correspond to the number of simulated samples

p

Required. The number of columns to simulate in the data. These correspond to the number of gene features

prop

Required. a value between 0 and 1, corresponding the probability value in the `rbinom()` adjustment of expression values.

adjustment

Optional. A matrix, vector, or 1D array, or missing. This corresponds to the x vaule in the `diag()` adjustment of the inverse covariance matrix. (Default = 10)

Value

A list containing a simulated network and the inverse covariance matrix of the simulated network.s