Runs vbsr with a 2Z cutoff for a gene across a matrix

This function wraps variable bays spike regression of a genes expression across a matrix of genes expressed in the same samples with a 2Z cutoff value.

sparrow2Z(
  y,
  x,
  fdr = NULL,
  ordering_mat = NULL,
  eps = 1e-06,
  exclude = NULL,
  add.intercept = TRUE,
  maxit = 10000,
  n_orderings = 10,
  family = "normal",
  scaling = TRUE,
  return_kl = TRUE,
  estimation_type = "BMA",
  bma_approximation = TRUE,
  screen = 1,
  post = 0.95,
  already_screened = 1,
  kl = 0.99,
  l0_path = NULL,
  cleanSolution = FALSE
)

Arguments

y

Required. response variable. Normally distributed errors for family="normal". For family="binomial" should be coded as a vector of 0's and 1's.

x

Required. Design matrix, an n x m matrix, with rows as observations.

fdr

Optional. FDR threshold cut off for edge determination. NULL results in a cutoff of 0.05 (Default = NULL)

ordering_mat

Optional. Optionally specified coordinate update ordering matrix. Must be in matrix form with columns as permutation vectors of length m, and there must be n_orderings columns. (Default = NULL)

eps

Optional. Tolerance used to determine convergence of the algorithm based on the lower bound. (Default = 1e-6)

exclude

Optional. An optional indicator vector of length m of 0's and 1's indicating whether to penalize a particular variable or not (0=penalize, 1=unpenalized) (Default = NULL)

add.intercept

Optional. A boolean variable indicating whether or not to include an unpenalized intercept variable. (Default = TRUE)

maxit

Optional. The maximum number of iterations to run the algorithm for a given solution to a penalized regression problem. (Default = 1e4)

n_orderings

Optional. The number of random starts used. (Default = 10)

family

Optional. The type of error model used. Currently supported modes are family="normal" and family="binomial". (Default = "normal")

scaling

Optional. The type of error model used. Currently supported modes are family="normal" and family="binomial" (Default = TRUE)

return_kl

Optional. A boolean variable indicating whether or not to return an analysis of the null distributed features in the data-set as a function of the penalty parameter. (Default = TRUE)

estimation_type

Optional. The type of estimation to perform based on the number of unique solution identified to the penalized regression problem. Valid values are estimation_type="BMA" and estimation_type="MAXIMAL" (Default = "BMA")

bma_approximation

Optional. A boolean variable indicating whether to compute a full correction to the z statistic. WARNING can make the algorithm very computationally intensive for highly multi-modal posterior surfaces. (Default = TRUE)

screen

Optional. P-value to do marginal screening. Default is to not do marginal prescreening (e.g marginal p-value of 1.0) (Default = 1.0)

post

Optional. Choice of penalty parameter such that a feature will have a posterior probability of 0.95 if it passes a Bonferroni correction in the multivariate model. Default is post=.95. More conservative approach would be post=0.5(Default =0.95)

already_screened

Optional. If features are already screened, the marginal p-value used for screening. (Default = 1.0)

kl

Optional. If features are already screened, the marginal p-value used for screening. (Default = 0.99)

l0_path

Optional. The path of penalty parameters to solve the spike regression problem. If post is specified, this is computed automatically. (Default =NULL)

cleanSolution

Optional. This parameter determines whether a given solution is further filtered using an unpenalized model. If cleanSolution=TRUE, then the features that are significant after a Bonferroni correction given the p-values from the vbsr regression model are then tested in an unpenalized linear regression model. The p-values and z-statistics are updated using the Wald test from the unpenalized linear regression model for the features that were selected.(Default =FALSE)

Value

A coexpression value