compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.41077501813
#> Epoch: Iteration #200 error is: 217.149632252989
#> Epoch: Iteration #300 error is: 217.280788969724
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  23.9699854   4.3168625  -4.7150988  16.0495267  10.7082106 -11.2447051
#>  [7] -31.1225455  -7.0596916 -24.0191628 -21.3960192 -15.2872787 -27.3844976
#> [13]  13.6438515  18.4888453  31.9489070   2.1172794  21.0107034  -2.3075092
#> [19]  -0.1618955 -17.7962425   7.7306987  27.5904994  -9.1671324 -13.3837952
#> [25]   6.3112917   8.9117161  11.6803424 -19.4331459
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775042512
#> Epoch: Iteration #200 error is: 217.132086910766
#> Epoch: Iteration #300 error is: 217.254362943016

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018261
#> Epoch: Iteration #200 error is: 217.105595761743
#> Epoch: Iteration #300 error is: 217.236219377924
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  23.9699854   4.3168625  -4.7150988  16.0495267  10.7082106 -11.2447051
#>  [7] -31.1225455  -7.0596916 -24.0191628 -21.3960192 -15.2872787 -27.3844976
#> [13]  13.6438515  18.4888453  31.9489070   2.1172794  21.0107034  -2.3075092
#> [19]  -0.1618955 -17.7962425   7.7306987  27.5904994  -9.1671324 -13.3837952
#> [25]   6.3112917   8.9117161  11.6803424 -19.4331459
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1]  22.5910506   6.0713111  -4.8456896   4.1371543  13.9986412 -11.2405401
#>  [7] -29.7338072  -7.1457204 -20.6608823 -22.7396358 -15.1695702 -26.0818708
#> [13]   8.2999314  17.7558711  25.9369211   1.8925191  20.0426132  -2.4595758
#> [19]  -0.3238957 -17.6442030  10.5687023  30.5042445  -9.2034691 -13.3683658
#> [25]  15.7590642  12.3863464  10.0090604 -19.3362053