compute and attach 'TL-tsne' layouts

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

'TL-tsne' or 'group-TL-tsne'

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network

NULL

Examples

library(metanetwork)
library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
# on angola dataset (metaweb)
data(meta_angola)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> attaching TL-tsne layout for metaweb_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019963
#> Epoch: Iteration #200 error is: 217.135975994572
#> Epoch: Iteration #300 error is: 217.266761513653
V(meta_angola$metaweb)$layout_beta0.05
#>  [1] -22.6466927  -5.9954139   4.7288334  -4.0963107 -12.3754322  15.0045533
#>  [7]  29.5063968   7.0362029  20.4884433  22.5466239  11.0469191  25.8767337
#> [13] -15.0576338 -17.6011403 -26.0127129  -1.9325290 -19.9281494   2.3645488
#> [19]   0.2607275  17.5048153  -9.4896210 -30.6643921   9.1024111  13.0297587
#> [25]  -8.0824473 -10.6234370 -13.1843642  19.1933086
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> attaching TL-tsne layout for X1986_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019007
#> Epoch: Iteration #200 error is: 217.152618246552
#> Epoch: Iteration #300 error is: 217.284834751365

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> attaching TL-tsne layout for metaweb_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775028761
#> Epoch: Iteration #200 error is: 217.120829152943
#> Epoch: Iteration #300 error is: 217.24150198486
V(meta_angola$metaweb)$layout_beta0.05
#>  [1] -22.6466927  -5.9954139   4.7288334  -4.0963107 -12.3754322  15.0045533
#>  [7]  29.5063968   7.0362029  20.4884433  22.5466239  11.0469191  25.8767337
#> [13] -15.0576338 -17.6011403 -26.0127129  -1.9325290 -19.9281494   2.3645488
#> [19]   0.2607275  17.5048153  -9.4896210 -30.6643921   9.1024111  13.0297587
#> [25]  -8.0824473 -10.6234370 -13.1843642  19.1933086
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1]  22.203665   5.835902  -4.763210   3.949462  15.120563 -14.670680
#>  [7] -28.878735  -7.002735 -19.927988 -21.943322 -10.861859 -25.264041
#> [13]   8.908386  17.319551  25.509564   1.767563  19.559296  -2.445811
#> [19]  -0.379091 -17.083074  10.601496  30.085631  -9.002714 -12.794654
#> [25]  11.761399  13.578049   7.542668 -18.725283