compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775028971
#> Epoch: Iteration #200 error is: 217.115033229025
#> Epoch: Iteration #300 error is: 217.222106024355
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  18.1071261   6.1400103  -4.8803563   4.1888391  14.3130686 -15.3072594
#>  [7] -27.1810723  -7.1896550 -30.9928088 -21.3117801 -11.2313556 -24.1488271
#> [13]   8.4239372  20.7752203  26.6727642   1.9009074  23.4770590  -2.4874064
#> [19]  -0.3362764 -17.8426484  10.7321862  31.3795801  -9.2842100 -13.2632044
#> [25]  16.0824935  12.5977881  10.1791583 -19.5132784
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775057795
#> Epoch: Iteration #200 error is: 217.121168312721
#> Epoch: Iteration #300 error is: 217.252841779112

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018859
#> Epoch: Iteration #200 error is: 217.114533727275
#> Epoch: Iteration #300 error is: 217.238377925835
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  18.1071261   6.1400103  -4.8803563   4.1888391  14.3130686 -15.3072594
#>  [7] -27.1810723  -7.1896550 -30.9928088 -21.3117801 -11.2313556 -24.1488271
#> [13]   8.4239372  20.7752203  26.6727642   1.9009074  23.4770590  -2.4874064
#> [19]  -0.3362764 -17.8426484  10.7321862  31.3795801  -9.2842100 -13.2632044
#> [25]  16.0824935  12.5977881  10.1791583 -19.5132784
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1]  25.3973980   8.0857050  -4.9322005   9.7991901   5.9009494 -15.1532120
#>  [7] -29.5791603  -7.2238284 -20.6178580 -22.6723856 -11.2073724 -25.9830632
#> [13]  19.4741676  13.1990598  22.0174156  -0.2567807  16.5779849  -2.5653169
#> [19]   2.0790441 -17.6542913   3.7937169  29.7236813  -9.2772594 -13.1984754
#> [25]  11.4746982  14.8880672  17.2592427 -19.3491170