compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019091
#> Epoch: Iteration #200 error is: 217.123696172265
#> Epoch: Iteration #300 error is: 217.235637126316
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  23.8052787   6.3500995  -4.9919766   4.3173525  14.7045216 -11.5750737
#>  [7] -27.8224298  -7.3285473 -31.6950657 -21.8876569 -15.7181934 -24.7342426
#> [13]   8.6624401  18.6676265  27.3249344   1.9849554  21.0783406  -2.5239404
#> [19]  -0.3179112 -18.2966381  11.0902724  32.0784544  -9.4512652 -13.7398486
#> [25]  16.5809706  12.9975238  10.4728128 -20.0327938
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775022257
#> Epoch: Iteration #200 error is: 217.111747473726
#> Epoch: Iteration #300 error is: 217.237510284764

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775029978
#> Epoch: Iteration #200 error is: 217.130303700915
#> Epoch: Iteration #300 error is: 217.249360734201
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  23.8052787   6.3500995  -4.9919766   4.3173525  14.7045216 -11.5750737
#>  [7] -27.8224298  -7.3285473 -31.6950657 -21.8876569 -15.7181934 -24.7342426
#> [13]   8.6624401  18.6676265  27.3249344   1.9849554  21.0783406  -2.5239404
#> [19]  -0.3179112 -18.2966381  11.0902724  32.0784544  -9.4512652 -13.7398486
#> [25]  16.5809706  12.9975238  10.4728128 -20.0327938
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1]  26.8355044   6.1920905  -4.8111888   4.2311100  14.2392482 -15.3088207
#>  [7] -31.0204894  -7.1705187 -24.0008308 -21.4150566 -11.2713428 -27.3434987
#> [13]  11.3415869  18.0833619  23.2147580   1.9947639  20.4218831  -2.4048987
#> [19]  -0.2487786 -17.8430636   9.6364648  31.4395459  -9.2600008 -13.3111979
#> [25]   8.3259281  15.9514154  12.9729019 -19.4708768