compute and attach 'TL-tsne' layouts

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

'TL-tsne' or 'group-TL-tsne'

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network

NULL

Examples

library(metanetwork)
library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
# on angola dataset (metaweb)
data(meta_angola)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> attaching TL-tsne layout for metaweb_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.41077501994
#> Epoch: Iteration #200 error is: 217.149420953341
#> Epoch: Iteration #300 error is: 217.284245372675
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  23.7935996   4.2547435  -4.8091953  15.9473790  10.6419574 -15.4021519
#>  [7] -30.4938900  -7.1765265 -21.1372741 -23.2704290 -11.3089739 -26.7087502
#> [13]  13.5957134  18.4217549  27.2905818   2.0487346  20.9027353  -2.3934579
#> [19]  -0.2380888 -17.9886292   7.6725326  32.0592207  -9.3169833 -13.3753589
#> [25]   6.2524380   8.8637578  11.6002348 -19.7256743
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> attaching TL-tsne layout for X1986_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019298
#> Epoch: Iteration #200 error is: 217.142414956809
#> Epoch: Iteration #300 error is: 217.261998411837

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> attaching TL-tsne layout for metaweb_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.41077502049
#> Epoch: Iteration #200 error is: 217.144656063022
#> Epoch: Iteration #300 error is: 217.271704410431
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  23.7935996   4.2547435  -4.8091953  15.9473790  10.6419574 -15.4021519
#>  [7] -30.4938900  -7.1765265 -21.1372741 -23.2704290 -11.3089739 -26.7087502
#> [13]  13.5957134  18.4217549  27.2905818   2.0487346  20.9027353  -2.3934579
#> [19]  -0.2380888 -17.9886292   7.6725326  32.0592207  -9.3169833 -13.3753589
#> [25]   6.2524380   8.8637578  11.6002348 -19.7256743
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1] -19.9790263  -5.9017299   4.5417117 -13.3972628  -8.4813035  10.8037542
#>  [7]  29.4565509   6.8010139  22.8125466  20.3748345  14.6515793  25.9597599
#> [13] -15.2545060 -23.0896881 -29.6573194  -2.0258958 -26.0620540   2.2156125
#> [19]   0.1135046  17.0302253  -4.1364685 -17.3967599   8.8047655  12.8535636
#> [25] -11.8350549 -10.0281660  -7.7389515  18.5647642