compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

Usage

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018073
#> Epoch: Iteration #200 error is: 217.117921401331
#> Epoch: Iteration #300 error is: 217.231830964565
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  18.1773513   6.3387259  -2.5156989   4.2557880  14.3539972 -11.2757448
#>  [7] -31.0358607  -4.8397862 -23.9839391 -21.3977528  -7.2747412 -27.2976576
#> [13]  10.7375621  20.8422871  26.6799094  -0.1177639  23.5155111 -15.3404072
#> [19]   2.2531012 -17.8379430   8.5288068  31.3281007  -9.1845559 -13.3597680
#> [25]  16.1837836  12.7842752   8.9878078 -19.5053881
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019159
#> Epoch: Iteration #200 error is: 217.125917916643
#> Epoch: Iteration #300 error is: 217.254374810136

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018624
#> Epoch: Iteration #200 error is: 217.122663306626
#> Epoch: Iteration #300 error is: 217.25243726917
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  18.1773513   6.3387259  -2.5156989   4.2557880  14.3539972 -11.2757448
#>  [7] -31.0358607  -4.8397862 -23.9839391 -21.3977528  -7.2747412 -27.2976576
#> [13]  10.7375621  20.8422871  26.6799094  -0.1177639  23.5155111 -15.3404072
#> [19]   2.2531012 -17.8379430   8.5288068  31.3281007  -9.1845559 -13.3597680
#> [25]  16.1837836  12.7842752   8.9878078 -19.5053881
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1]  23.0082322   6.1008829  -4.8228874   4.1517190  14.1356611 -15.1547581
#>  [7] -30.6678522  -7.1222101 -23.6977066 -21.1254352 -11.1428182 -26.9967478
#> [13]  11.2206509  17.9652923  26.3786787   1.9252790  20.2885132  -2.4267180
#> [19]  -0.3017448 -17.6393118   9.5648710  31.0035875  -9.2247746 -13.1686098
#> [25]   8.2482646  15.8131111  12.9148870 -19.2280556