compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018387
#> Epoch: Iteration #200 error is: 217.146392309045
#> Epoch: Iteration #300 error is: 217.270105926774
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  26.7094949   4.3418937  -4.5639012  15.5718066   9.8375934 -10.9018324
#>  [7] -30.0468785  -6.8644428 -23.1548481 -20.6409603 -14.8100942 -26.4229471
#> [13]  13.4450284  17.7882032  20.4208288   2.0997188  23.2256482  -2.2252601
#> [19]  -0.1044616 -17.1901792   6.6276990  31.3296591  -8.8609660 -12.9432425
#> [25]  11.3900706   8.4000600   6.2949010 -18.7525915
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775023296
#> Epoch: Iteration #200 error is: 217.133724522158
#> Epoch: Iteration #300 error is: 217.263308208265

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018594
#> Epoch: Iteration #200 error is: 217.113805166448
#> Epoch: Iteration #300 error is: 217.227942491408
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  26.7094949   4.3418937  -4.5639012  15.5718066   9.8375934 -10.9018324
#>  [7] -30.0468785  -6.8644428 -23.1548481 -20.6409603 -14.8100942 -26.4229471
#> [13]  13.4450284  17.7882032  20.4208288   2.0997188  23.2256482  -2.2252601
#> [19]  -0.1044616 -17.1901792   6.6276990  31.3296591  -8.8609660 -12.9432425
#> [25]  11.3900706   8.4000600   6.2949010 -18.7525915
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1]  18.084955   6.126960  -4.896277   4.151512  14.196777 -15.257226
#>  [7] -30.885350  -7.198296 -23.882019 -21.302699 -11.223672 -27.188342
#> [13]   8.381810  23.801721  30.659052   1.879434  26.903713  -2.503626
#> [19]  -0.355309 -17.772538  10.692108  20.749880  -9.279845 -13.241345
#> [25]  16.046372  12.542283  10.163760 -19.393793