compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

Usage

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018853
#> Epoch: Iteration #200 error is: 217.120073096718
#> Epoch: Iteration #300 error is: 217.235823826717
V(meta_angola$metaweb)$layout_beta0.05
#>  [1] -23.7924607  -6.3522508   4.9780791  -4.3303445 -14.6968172  11.5400695
#>  [7]  31.7070017  15.7538755  24.5943293  21.9491981   7.4542280  27.9692360
#> [13]  -8.6515426 -18.6691574 -27.3019077  -2.0103163 -21.0765124   2.5205443
#> [19]   0.2865085  18.2816662 -11.0884443 -32.0756950   9.3709157  13.6814619
#> [25] -16.5714690 -12.9766905 -10.4692298  19.9757246
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019562
#> Epoch: Iteration #200 error is: 217.118675041551
#> Epoch: Iteration #300 error is: 217.250517581737

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775020073
#> Epoch: Iteration #200 error is: 217.137478983339
#> Epoch: Iteration #300 error is: 217.269600252919
V(meta_angola$metaweb)$layout_beta0.05
#>  [1] -23.7924607  -6.3522508   4.9780791  -4.3303445 -14.6968172  11.5400695
#>  [7]  31.7070017  15.7538755  24.5943293  21.9491981   7.4542280  27.9692360
#> [13]  -8.6515426 -18.6691574 -27.3019077  -2.0103163 -21.0765124   2.5205443
#> [19]   0.2865085  18.2816662 -11.0884443 -32.0756950   9.3709157  13.6814619
#> [25] -16.5714690 -12.9766905 -10.4692298  19.9757246
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1] -23.460767  -6.229209   4.926822  -4.229971 -14.419353  15.645064
#>  [7]  24.192622   7.295545  30.593278  18.223351  11.460590  21.792682
#> [13] -11.472265 -18.315172 -26.888245  -1.962436 -20.701617   2.490973
#> [19]   0.312691  26.775006  -9.743265 -31.556530   9.436225  13.547302
#> [25]  -8.395388 -16.164080 -13.147127  19.993272