compute and attach metanetwork layouts

Method to compute 'TL-tsne' and 'group-TL-tsne' layouts and save it as node attributes of the focal network.

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network, default is 0.1

mode

'TL-tsne' or 'group-TL-tsne', default is 'TL-tsne'.

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

res

resolution for the 'group-TL-tsne' layout

group_layout.config

configuration list for mode 'group-TL-tsne', default is group_layout.default

Value

'metanetwork' object with layout added as node attribute of the considered network

NULL

Details

The 'TL-tsne' layout is a diffusion based layout algorithm specifically designed for trophic networks. In metanetwork, first axis is the trophic level (see compute_TL method) whereas the second axis is computed using a diffusion graph kernel (Kondor & Lafferty 2002) and tsne dimension reduction algorithm to (see van der Maaten & Hinton (2008) and 'tsne' R package). Let \(A\) be the adjacency matrix of the considered network and \(D\) its degree diagonal matrix. The Laplacian matrix of the symmetrised network is defined by:

$$L = D - A - t(A)$$

The diffusion graph kernel is:

$$K = exp(-beta*L)$$

It is a similarity matrix between nodes according to a diffusion process. beta is the diffusion constant,it must be provided by the user. beta parameter influences the layout by grouping together similar paths (see pyramid vignette). Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes.

The 'group-TL-tsne' layout is a variation of 'TL-tsne layout. For a focal network, it mixes 'TL-tsne' layout at the desired aggregated level with the layout_with_graphopt function from igraph. It clusters nodes belonging to the same group. 'group-TL-tsne' layout is recommended for large networks since you only need to compute 'TL-tsne' at the aggregated network that is much smaller than the focal network. group_layout.config allows controlling the overall size of the groups.

References

Kondor, R. I., & Lafferty, J. (2002, July). Diffusion kernels on graphs and other discrete structures. In Proceedings of the 19th international conference on machine learning (Vol. 2002, pp. 315-322). Van der Maaten, L., & Hinton, G. (2008). Visualizing data using t-SNE. Journal of machine learning research, 9(11).

See also

Examples

library(metanetwork)
library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
# on angola dataset (metaweb)
data("meta_angola")
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> attaching TL-tsne layout for metaweb_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775031149
#> Epoch: Iteration #200 error is: 217.116377183544
#> Epoch: Iteration #300 error is: 217.231520307094
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  22.8495241   6.1297970  -4.8046146   4.1672160  14.1405500 -15.0741868
#>  [7] -26.7246535  -7.1150381 -30.4976666 -20.9656886 -11.0959335 -23.7441631
#> [13]   8.3489506  17.9102151  30.5433073   1.9188347  20.1973037  -2.4457793
#> [19]  -0.3067904 -17.5617698  10.6755367  26.3500879  -9.1404784 -13.0635720
#> [25]  15.9139279  12.4984214  10.0999416 -19.2032791
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> attaching TL-tsne layout for X1986_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775023533
#> Epoch: Iteration #200 error is: 217.127641213909
#> Epoch: Iteration #300 error is: 217.24880709912

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> attaching TL-tsne layout for metaweb_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018467
#> Epoch: Iteration #200 error is: 217.133590526988
#> Epoch: Iteration #300 error is: 217.250797410851
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  22.8495241   6.1297970  -4.8046146   4.1672160  14.1405500 -15.0741868
#>  [7] -26.7246535  -7.1150381 -30.4976666 -20.9656886 -11.0959335 -23.7441631
#> [13]   8.3489506  17.9102151  30.5433073   1.9188347  20.1973037  -2.4457793
#> [19]  -0.3067904 -17.5617698  10.6755367  26.3500879  -9.1404784 -13.0635720
#> [25]  15.9139279  12.4984214  10.0999416 -19.2032791
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1] -25.7601141  -5.9140019   4.4937617  -4.0749999 -12.1745072  14.5118639
#>  [7]  25.9357644   6.7351060  29.6402292  20.2490437  10.6269943  23.0052901
#> [13] -14.7400183 -17.2276858 -22.1875484  -1.9694070 -19.4721301   2.1968011
#> [19]   0.1599815  16.9588367  -9.3275718 -30.3123715   8.7320889  12.5561849
#> [25]  -7.9485814 -10.4061806 -12.8263289  18.5395006

# attaching 'group-TL-tsne' layout on norway dataset
data("meta_norway")
## attach 'TL-tsne layout at the desired aggregated level
meta_norway = append_agg_nets(meta_norway) %>% 
              compute_TL()
meta_norway = attach_layout(meta_norway,meta_norway$metaweb_trophic_class,beta = 0.1)
#> attaching TL-tsne layout for metaweb_trophic_class
#> 
#>                   beta = 0.1
#> beta = 0.1
#> Epoch: Iteration #100 error is: 275.793322650735
#> Epoch: Iteration #200 error is: 65.6945503852518
#> Epoch: Iteration #300 error is: 65.7135649472163
## attach 'group-TL-tsne layout'
meta_norway = attach_layout(meta_norway,mode = "group-TL-tsne",res = "trophic_class",beta = 0.1)
#> attaching group-TL-tsne layout for metaweb at resolution: trophic_class
#> 
#>                   beta = 0.1