compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019633
#> Epoch: Iteration #200 error is: 217.130071884779
#> Epoch: Iteration #300 error is: 217.255339701972
V(meta_angola$metaweb)$layout_beta0.05
#>  [1] -25.9925389  -6.3090973   4.7575262  -4.2817842 -11.4694875  15.1800342
#>  [7]  27.0121851  11.0803670  30.8285711  21.1594613   7.1488670  23.9693944
#> [13] -19.8105219 -14.7396958 -22.4921385  -2.0435522 -17.1416070   2.3584345
#> [19]   0.2016443  17.7141103  -8.2285119 -30.4613401   8.9931231  13.1574010
#> [25]  -9.7442885 -13.0689334 -17.1543200  19.3766976
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018596
#> Epoch: Iteration #200 error is: 217.131651445781
#> Epoch: Iteration #300 error is: 217.255984770807

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018886
#> Epoch: Iteration #200 error is: 217.111878572859
#> Epoch: Iteration #300 error is: 217.231606992088
V(meta_angola$metaweb)$layout_beta0.05
#>  [1] -25.9925389  -6.3090973   4.7575262  -4.2817842 -11.4694875  15.1800342
#>  [7]  27.0121851  11.0803670  30.8285711  21.1594613   7.1488670  23.9693944
#> [13] -19.8105219 -14.7396958 -22.4921385  -2.0435522 -17.1416070   2.3584345
#> [19]   0.2016443  17.7141103  -8.2285119 -30.4613401   8.9931231  13.1574010
#> [25]  -9.7442885 -13.0689334 -17.1543200  19.3766976
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1]  23.8416214   6.3354896  -5.0216460   4.3006744  14.7474918 -15.7724323
#>  [7] -31.7298590  -7.3823516 -24.5897226 -21.9445256 -11.5743498 -27.9557949
#> [13]   8.6481021  18.7237242  27.3464049   1.9662366  21.1201561  -2.5521968
#> [19]  -0.3360965 -18.2953667  11.1062248  32.1210537  -9.5535328 -13.6310758
#> [25]  16.6220510  13.0106940  10.4719562 -20.0229307