compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775028481
#> Epoch: Iteration #200 error is: 217.115652358542
#> Epoch: Iteration #300 error is: 217.228459424264
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  23.0535023   6.1723524  -4.8337090   4.2017966  14.2832629 -15.2370318
#>  [7] -27.0567486  -7.1375125 -30.8729797 -21.2048186 -11.1700483 -24.0155799
#> [13]   8.4253706  18.1076752  26.4557812   1.9293647  20.4278932  -2.4519991
#> [19]  -0.3094649 -17.7489262  10.7864135  31.1009096  -9.2037199 -13.1766096
#> [25]  16.0952760  12.6186354  10.1794577 -19.4185433
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.41077501913
#> Epoch: Iteration #200 error is: 217.131641496299
#> Epoch: Iteration #300 error is: 217.261674021806

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018246
#> Epoch: Iteration #200 error is: 217.144852609124
#> Epoch: Iteration #300 error is: 217.267738853077
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  23.0535023   6.1723524  -4.8337090   4.2017966  14.2832629 -15.2370318
#>  [7] -27.0567486  -7.1375125 -30.8729797 -21.2048186 -11.1700483 -24.0155799
#> [13]   8.4253706  18.1076752  26.4557812   1.9293647  20.4278932  -2.4519991
#> [19]  -0.3094649 -17.7489262  10.7864135  31.1009096  -9.2037199 -13.1766096
#> [25]  16.0952760  12.6186354  10.1794577 -19.4185433
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1] -23.5658447  -4.3125781   4.6128528 -15.8558232 -10.6216305  15.0975507
#>  [7]  31.1662455   6.9466978  26.9078528  23.8130325  11.0190396  21.0642476
#> [13] -13.4916176 -18.2539000 -27.0377481  -2.1385212 -20.7262249   2.2340762
#> [19]   0.1159159  17.6820321  -7.6763019 -31.7790186   9.0469781  13.0515898
#> [25]  -6.2883780  -8.8406356 -11.5428840  19.3729950