compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775017876
#> Epoch: Iteration #200 error is: 217.142422514677
#> Epoch: Iteration #300 error is: 217.271403701004
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  23.5576492   4.3118091  -4.6424547  15.8215128  10.6050564 -15.1008806
#>  [7] -30.7129046  -6.9834326 -23.7301147 -21.1905288 -11.0526424 -27.0237677
#> [13]  13.4601385  18.2259501  27.0322028   2.1270496  20.7075505  -2.2578840
#> [19]  -0.1273928 -17.6732364   7.6696575  31.7643774  -9.0819999 -13.0817994
#> [25]   6.2750980   8.8219716  11.5037887 -19.2247737
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775021664
#> Epoch: Iteration #200 error is: 217.127504715553
#> Epoch: Iteration #300 error is: 217.259505711904

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775071714
#> Epoch: Iteration #200 error is: 217.119591617667
#> Epoch: Iteration #300 error is: 217.235842351835
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  23.5576492   4.3118091  -4.6424547  15.8215128  10.6050564 -15.1008806
#>  [7] -30.7129046  -6.9834326 -23.7301147 -21.1905288 -11.0526424 -27.0237677
#> [13]  13.4601385  18.2259501  27.0322028   2.1270496  20.7075505  -2.2578840
#> [19]  -0.1273928 -17.6732364   7.6696575  31.7643774  -9.0819999 -13.0817994
#> [25]   6.2750980   8.8219716  11.5037887 -19.2247737
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1] -23.7213701  -6.3432703   4.9812673  -4.3127950 -14.7012708  15.6583743
#>  [7]  31.6508980   7.3425421  24.4688979  21.8320367  11.5008600  27.8711004
#> [13]  -8.6524044 -18.6151580 -31.5760519  -1.9822519 -20.9817454   2.5257411
#> [19]   0.3157512  18.2138494 -11.0769840 -27.2965480   9.4964827  13.5679749
#> [25] -16.5481832 -12.9802879 -10.4789901  19.8415349