compute and attach 'TL-tsne' layouts

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

'TL-tsne' or 'group-TL-tsne'

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network

NULL

Examples

library(metanetwork)
library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
# on angola dataset (metaweb)
data(meta_angola)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> attaching TL-tsne layout for metaweb_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775021495
#> Epoch: Iteration #200 error is: 217.111574612431
#> Epoch: Iteration #300 error is: 217.237758016704
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  25.2249316   8.0615050  -2.4926979   9.7479865   5.8863778  -7.1020079
#>  [7] -30.0124939  -4.7480767 -23.1769297 -20.6726498 -14.8262272 -26.4111845
#> [13]  19.3760468  13.1172805  21.8917553  -0.1786442  16.5058257 -10.9311575
#> [19]   2.1229883 -17.2433044   3.8064580  29.4785125  -9.1017015 -13.0067882
#> [25]  11.4541737  14.8454082  17.1788015 -18.7941880
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> attaching TL-tsne layout for X1986_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019902
#> Epoch: Iteration #200 error is: 217.139262576185
#> Epoch: Iteration #300 error is: 217.25473990688

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> attaching TL-tsne layout for metaweb_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775017929
#> Epoch: Iteration #200 error is: 217.137294607864
#> Epoch: Iteration #300 error is: 217.281796779292
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  25.2249316   8.0615050  -2.4926979   9.7479865   5.8863778  -7.1020079
#>  [7] -30.0124939  -4.7480767 -23.1769297 -20.6726498 -14.8262272 -26.4111845
#> [13]  19.3760468  13.1172805  21.8917553  -0.1786442  16.5058257 -10.9311575
#> [19]   2.1229883 -17.2433044   3.8064580  29.4785125  -9.1017015 -13.0067882
#> [25]  11.4541737  14.8454082  17.1788015 -18.7941880
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1] -22.0813652  -5.7078579   4.5878916  -3.8937471 -11.7603726  10.7404172
#>  [7]  19.4591616   6.8155660  24.0172627  21.4837735  14.5776454  27.3244374
#> [13] -14.2559563 -16.5798940 -29.0973353  -1.8076695 -25.4658197   2.3088085
#> [19]   0.2970368  17.0574957  -9.0138002 -19.1301767   8.7662404  12.7559473
#> [25]  -7.7187851 -10.1038973 -12.4540923  18.8790851