compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775021776
#> Epoch: Iteration #200 error is: 217.128742481594
#> Epoch: Iteration #300 error is: 217.248673456097
V(meta_angola$metaweb)$layout_beta0.05
#>  [1] -23.5515911  -6.2326519   4.8986744  -4.2274270 -14.6945519  15.4149015
#>  [7]  31.2262090   7.2409908  24.1472625  21.5295844  11.3332196  27.4999302
#> [13]  -9.5759340 -18.3265343 -31.3792413  -1.9397677 -20.7400097   2.4836410
#> [19]   0.3084363  17.9509491 -11.3417974 -27.0967329   9.3466670  13.3767532
#> [25] -12.9149093 -16.2948136  -8.0296688  19.5884120
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018137
#> Epoch: Iteration #200 error is: 217.143150470533
#> Epoch: Iteration #300 error is: 217.271766957718

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019988
#> Epoch: Iteration #200 error is: 217.139200530909
#> Epoch: Iteration #300 error is: 217.26977427702
V(meta_angola$metaweb)$layout_beta0.05
#>  [1] -23.5515911  -6.2326519   4.8986744  -4.2274270 -14.6945519  15.4149015
#>  [7]  31.2262090   7.2409908  24.1472625  21.5295844  11.3332196  27.4999302
#> [13]  -9.5759340 -18.3265343 -31.3792413  -1.9397677 -20.7400097   2.4836410
#> [19]   0.3084363  17.9509491 -11.3417974 -27.0967329   9.3466670  13.3767532
#> [25] -12.9149093 -16.2948136  -8.0296688  19.5884120
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1]  23.6351670   6.2040859  -4.9237521   4.2052959  14.4478610 -11.4916010
#>  [7] -24.3457239  -7.2782771 -30.6021984 -27.1196151 -15.6317419 -21.8600171
#> [13]  11.4814586  18.3735558  31.4825876   1.9386766  20.7893168  -2.5025793
#> [19]  -0.3352567 -18.2120628   9.7383264  27.1945871  -9.3715481 -13.6443755
#> [25]   8.3714559  16.2184594  13.1635055 -19.9255906