compute and attach 'TL-tsne' layout

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

only 'TL-tsne' is supported for this function

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network NULL

Examples

library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019682
#> Epoch: Iteration #200 error is: 217.12039627624
#> Epoch: Iteration #300 error is: 217.258520669387
V(meta_angola$metaweb)$layout_beta0.05
#>  [1] -25.21541703  -5.88255217   4.64476720  -3.98169643 -12.06728541
#>  [6]  14.62067716  28.68986159   6.88435037  19.86173476  21.86623346
#> [11]  10.77391762  25.12727789 -14.71067527 -19.61246381 -29.13179573
#> [16]   0.09511713 -22.15256605   2.32850025  -2.11209611  17.01456460
#> [21]  -9.23073402 -16.91095269   8.89554269  12.68052089  -7.92466465
#> [26] -10.37640415 -12.82672632  18.65296424
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018164
#> Epoch: Iteration #200 error is: 217.112616755559
#> Epoch: Iteration #300 error is: 217.231233731001

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018656
#> Epoch: Iteration #200 error is: 217.129399239016
#> Epoch: Iteration #300 error is: 217.243510216642
V(meta_angola$metaweb)$layout_beta0.05
#>  [1] -25.21541703  -5.88255217   4.64476720  -3.98169643 -12.06728541
#>  [6]  14.62067716  28.68986159   6.88435037  19.86173476  21.86623346
#> [11]  10.77391762  25.12727789 -14.71067527 -19.61246381 -29.13179573
#> [16]   0.09511713 -22.15256605   2.32850025  -2.11209611  17.01456460
#> [21]  -9.23073402 -16.91095269   8.89554269  12.68052089  -7.92466465
#> [26] -10.37640415 -12.82672632  18.65296424
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1] -23.9214665  -6.0611463   4.9796867  -4.1477482 -16.2672988  15.5936384
#>  [7]  31.6737329   7.3417094  24.4595049  21.7964090  11.4736116  27.8662484
#> [13] -18.5616677  -9.9319084 -27.4171642  -1.8780777 -21.0843257   2.5621578
#> [19]   0.3962942  18.1517313 -13.7140379 -32.1776963   9.4658456  13.5497529
#> [25]  -8.1982645 -11.6922429 -14.0770582  19.8197801