compute and attach 'TL-tsne' layouts

Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute layout_beta_VALUE. If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes

Usage

attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

# S3 method for metanetwork
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

Arguments

metanetwork

object of class 'metanetwork'

g

character indicating the name of the network for which the 'TL-tsne' layout is computed, default is 'metaweb'

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the squeezing of the network

mode

'TL-tsne' or 'group-TL-tsne'

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

'metanetwork' object with layout added as node attribute of the considered network

NULL

Examples

library(metanetwork)
library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
# on angola dataset (metaweb)
data(meta_angola)
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> attaching TL-tsne layout for metaweb_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775019024
#> Epoch: Iteration #200 error is: 217.126864209189
#> Epoch: Iteration #300 error is: 217.24280219393
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  30.9785703   6.1278969  -5.0204093   4.0573611   9.9936091 -11.5372962
#>  [7] -27.6262598  -7.3475551 -31.4800563 -21.6829368 -15.6298002 -24.5487497
#> [13]  20.6814934  13.6401424  26.6272109  -0.2816853  23.5892366  -2.6165565
#> [19]   2.0479329 -18.1533435   7.9225422  17.3559587  -9.4152954 -13.6717796
#> [25]  11.8837870  15.4484838  18.5122075 -19.8547089
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)
#> attaching TL-tsne layout for X1986_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775018913
#> Epoch: Iteration #200 error is: 217.134328202294
#> Epoch: Iteration #300 error is: 217.264578711775

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
#> attaching TL-tsne layout for metaweb_Species
#> 
#>                   beta = 0.05
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775024787
#> Epoch: Iteration #200 error is: 217.109580616457
#> Epoch: Iteration #300 error is: 217.228908019609
V(meta_angola$metaweb)$layout_beta0.05
#>  [1]  30.9785703   6.1278969  -5.0204093   4.0573611   9.9936091 -11.5372962
#>  [7] -27.6262598  -7.3475551 -31.4800563 -21.6829368 -15.6298002 -24.5487497
#> [13]  20.6814934  13.6401424  26.6272109  -0.2816853  23.5892366  -2.6165565
#> [19]   2.0479329 -18.1533435   7.9225422  17.3559587  -9.4152954 -13.6717796
#> [25]  11.8837870  15.4484838  18.5122075 -19.8547089
V(meta_angola$metaweb)$layout_beta0.05_1
#>  [1] -17.3976876  -5.9261310   4.8402129  -4.0208954 -13.7203849  14.9932741
#>  [7]  29.2965295  11.0169220  20.3836020  22.4255014   7.1731891  25.7188397
#> [13]  -8.1205685 -19.9762043 -25.6892938  -1.8067716 -22.5957567   2.4891729
#> [19]   0.3788765  17.4415273 -10.3388262 -30.2707894   8.9764164  13.0468496
#> [25] -15.4826580 -12.1342239  -9.8290192  19.1282972