vismetaNetwork

Function that provides network dynamic representation (using 'visNetwork') from a 'metanetwork' object with a layout based on a diffusion kernel

vismetaNetwork(
  metanetwork,
  g = NULL,
  beta = 0.1,
  legend = NULL,
  mode = "TL-tsne",
  edge_thrs = NULL,
  layout_metaweb = F,
  nrep_ly = 1,
  flip_coords = F,
  diff_plot_bool = F,
  x_y_range = c(100, 100),
  visNetwork.config = visNetwork.default,
  TL_tsne.config = TL_tsne.default
)

Arguments

metanetwork

object of class metanetwork

g

network (igraph object) to represent, default is metaweb

beta

the diffusion parameter of the diffusion kernel, a positive scalar controlling the vertical squeezing of the network

legend

resolution for the legend, legend resolution must be a coarser resolution than the resolution of g, default is NULL

mode

mode used for layout, 'TL-tsne' for trophic level t-sne and 'TL-kpco' for trophic level kernel based pco. Default is 'TL-tsne'

edge_thrs

if non-null, a numeric (between 0 and 1) indicating an edge threshold for the representation

layout_metaweb

a boolean indicating whether the layout of the metaweb should be used to represent the network to use metaweb layout = T, you need first to compute metaweb layout for this beta value using attach_layout()

nrep_ly

If several layouts for this beta value are attached to the metaweb (if layout_metaweb = T), index of the layout to use, see attach_layout()

flip_coords

a boolean indicating wheter coordinates should be flipped. In that case, y-axis is the trophic level and x-axis is the layout axis

diff_plot_bool

boolean, do not edit by hand

x_y_range

a two dimension numeric vector, indicating dilatation of x,y axis

visNetwork.config

configuration list for visNetwork representation, default is visNetwork.default

TL_tsne.config

configuration list for mode 'TL-tsne', default is TL_tsne.default

Value

object of class 'visNetwork'

Examples

library(metanetwork)
library(igraph)

#on angola dataset
meta_angola = attach_layout(meta_angola, beta = 0.05)
#> beta = 0.05
#> Epoch: Iteration #100 error is: 907.410775025575
#> Epoch: Iteration #200 error is: 217.133483692899
#> Epoch: Iteration #300 error is: 217.254591034061
vismetaNetwork(meta_angola, beta = 0.05)
#> x_max = 1259.92651522773
#> y_max = 361.50254682172
#> Error in x$width %||% settings$fig.width * settings$dpi: non-numeric argument to binary operator
# flipping coordinates and controlling x_y_range
vismetaNetwork(meta_angola, beta = 0.05,flip_coords = TRUE,x_y_range = c(50,200))
#> x_max = 1446.01018728688
#> y_max = 314.981628806931
#> Error in x$width %||% settings$fig.width * settings$dpi: non-numeric argument to binary operator
#adding a legend
vismetaNetwork(meta_angola,legend = "Phylum", beta = 0.05)
#> x_max = 1259.92651522773
#> y_max = 361.50254682172
#> Error in x$width %||% settings$fig.width * settings$dpi: non-numeric argument to binary operator
#use metaweb layout to represent a local network
vismetaNetwork(metanetwork = meta_angola,g = meta_angola$X2003,legend = "Phylum", beta = 0.05,
layout_metaweb=TRUE)
#> x_max = 1259.92651522773
#> y_max = 361.50254682172
#> Error in x$width %||% settings$fig.width * settings$dpi: non-numeric argument to binary operator

#custom visNetwork parameters
visNetwork.custom = visNetwork.default
visNetwork.custom$edge.size = 5
visNetwork.custom$label = FALSE

vismetaNetwork(meta_angola,legend = "Phylum", beta = 0.05,
visNetwork.config = visNetwork.custom)
#> x_max = 1259.92651522773
#> y_max = 361.50254682172
#> Error in x$width %||% settings$fig.width * settings$dpi: non-numeric argument to binary operator